+++ packages/interceptor/apply_forwarded_headers.test.ts
@@ -20,3 +20,42 @@ Deno.test("x-forwarded", () => {
   assertEquals(adjustedReq.method, "PUT");
   assertEquals(adjustedReq.body, incomingBody);
 });
+
+Deno.test("x-forwarded-host uses the leftmost value of a comma-joined list", () => {
+  const incomingReq = new Request("http://localhost:8000/foo", {
+    headers: {
+      "X-Forwarded-Proto": "https",
+      "X-Forwarded-Host": "somewhere.cool, browser-host.example.com",
+      "X-Forwarded-Port": "443",
+    },
+  });
+  const adjustedReq = applyForwardedHeaders(incomingReq);
+
+  assertEquals(adjustedReq.url, "https://somewhere.cool/foo");
+});
+
+Deno.test("x-forwarded-port does not override a port already in x-forwarded-host", () => {
+  const incomingReq = new Request("http://localhost:8000/foo", {
+    headers: {
+      "X-Forwarded-Proto": "https",
+      "X-Forwarded-Host": "somewhere.cool:8443",
+      "X-Forwarded-Port": "80",
+    },
+  });
+  const adjustedReq = applyForwardedHeaders(incomingReq);
+
+  assertEquals(adjustedReq.url, "https://somewhere.cool:8443/foo");
+});
+
+Deno.test("x-forwarded-port fills in when the forwarded host carries no port", () => {
+  const incomingReq = new Request("http://localhost:8000/foo", {
+    headers: {
+      "X-Forwarded-Proto": "https",
+      "X-Forwarded-Host": "somewhere.cool",
+      "X-Forwarded-Port": "8443",
+    },
+  });
+  const adjustedReq = applyForwardedHeaders(incomingReq);
+
+  assertEquals(adjustedReq.url, "https://somewhere.cool:8443/foo");
+});